Day 8, Part 2: Complex, but quite fun
Day 8 took way longer than it should have.  I was both fighting the Rust
compiler (and my lack of knowledge of Rust) and also the approach I took
to solve it.  I was very keen to do something that was performant (no
idea why, just took my fancy), and thus I went for a 'solver' that
actually coded how to solve the 'equations' for mapping a segment to the
randomised segment in each line.  This meant, for me, writing a bunch
structs and their implementations to make it easier to reason about.  It
could be much simpler if I didn't create custom structs for the data
structures.  Ho Hum.  It is pretty fast though, as it uses bit maps to
encode the 7-segment displays and also to 'solve' the equations.