Day 12 part 2
I broke part 1 slightly, which is annoying, but as I got the across/down
the wrong way around in the first part, I needed to correct it for the
2nd part.  Not sure why that broke the count, though.  hey/ho.

The 2nd part is fixed though and produces the correct puzzle output.
Also, I got to experiment with moving closures around and using Iterator
traits and lifetimes, which was fun!