Initial Day 18 part 1, almost complete.
This solution, which gets the right answer in terms of Pair calcuations,
does too much work during the explode, in that it recreates the
hierarchy of Pairs.  The split does mutate-in-place which is much more
efficient.  I'm going to change explode to mutate-in-place to make it
more efficient, and it's fun to explore these kinds of optimisations.