Day 18, part 1, fully mutable version.
This version uses a mutable explode (as well as the previously mutable
split).  This means that it doesn't recreate the complete tree, but just
mutates the bits it needs to. Consequently, it's faster and uses less
memory.  Yay!

Still missing the magnitude of the Pair to actually solve the exercise.